Create problem6.py
Problem 6: Determine the smallest rotated string
Given a string of length n and an integer k. The string can be manipulated by taking one of the first k letters and moving it to the end of the string.
Write a program to determine the lexicographically smallest string that can be created after an unlimited number of moves. 
For example, suppose we are give the string daily and k = 1. The best we can create in this case is ailyd.